Components Refactor
Refactored components in app.js to reside in separate files under components folder
