+++ src/components/Header.js
@@ -1,0 +1,16 @@
+import React from 'react';
+
+const Header = props => {
+	return (
+		<div>
+			<h1>{props.title}</h1>
+			{props.subtitle && <h2>{props.subtitle}</h2>}
+		</div>
+	);
+};
+
+Header.defaultProps = {
+	title: 'Indecision'
+};
+
+export default Header;
