+++ src/components/Option.js
@@ -1,0 +1,14 @@
+import React from 'react';
+
+const Option = props => {
+	return (
+		<div>
+			{props.optionText}
+			<button onClick={e => props.handleDeleteOption(props.optionText)}>
+				Remove
+			</button>
+		</div>
+	);
+};
+
+export default Option;
